+++ docs/changelog.md
@@ -9,6 +9,7 @@
 - [**v0.19.2**]
     - Adds a `border` field to the constellation model
     - Modifies the `line` function to accept a Shapely geometry OR a list of coordinates
+    - Fixes bug with plotting true apparent size of solar system objects
 
 ---